+++ pool/src/lib.rs
@@ -6,6 +6,11 @@ pub mod msg;
 pub mod response;
 pub mod state;
 pub mod oracle;
+pub mod liquidity;
+pub mod generic_helpers;
+pub mod query;
+pub mod liquidity_helpers;
+pub mod swap_helper;
 
 #[cfg(test)]
 mod testing;
