+++ pool/src/testing/tests.rs
@@ -2,15 +2,13 @@
 use std::str::FromStr;
 
 use cosmwasm_std::{
-    testing::{mock_dependencies, mock_env, mock_info, MockApi, MockQuerier, MockStorage, MOCK_CONTRACT_ADDR}, to_json_binary, Addr, BankMsg, Binary, Coin, ContractResult, CosmosMsg, Decimal, Order, OwnedDeps, SystemError, SystemResult, Timestamp, Uint128, WasmMsg, WasmQuery
+    testing::{mock_dependencies, mock_env, mock_info, MockApi, MockQuerier, MockStorage, MOCK_CONTRACT_ADDR}, 
+    to_json_binary, Addr, BankMsg, Binary, Coin, ContractResult, CosmosMsg, Decimal, Order, OwnedDeps, SystemError, 
+    SystemResult, Timestamp, Uint128, WasmMsg, WasmQuery
 };
 use cw20::Cw20ReceiveMsg;
 use cw721::OwnerOfResponse;
-use crate::{asset::PairType, 
-    contract::{calculate_fee_multiplier, execute, execute_add_to_position, execute_collect_fees, execute_deposit_liquidity, execute_remove_liquidity, execute_swap_cw20, instantiate, trigger_threshold_payout}, 
-    msg::{Cw20HookMsg, FeeInfo, PoolInstantiateMsg}, 
-    oracle::PriceResponse, 
-    state::{CommitInfo, OracleInfo, PoolFeeState, PoolInfo, PoolSpecs, PoolState, ThresholdPayout, COMMITSTATUS, 
+use crate::{asset::PairType, contract::{execute, execute_swap_cw20, instantiate}, generic_helpers::trigger_threshold_payout, liquidity::{execute_add_to_position, execute_collect_fees, execute_deposit_liquidity, execute_remove_liquidity}, liquidity_helpers::calculate_fee_multiplier, msg::{Cw20HookMsg, FeeInfo, PoolInstantiateMsg}, oracle::PriceResponse, state::{CommitInfo, OracleInfo, PoolFeeState, PoolInfo, PoolSpecs, PoolState, ThresholdPayout, COMMITSTATUS, 
         COMMIT_CONFIG, FEEINFO, LIQUIDITY_POSITIONS, NATIVE_RAISED, ORACLE_INFO, POOL_INFO, POOL_SPECS, THRESHOLD_PAYOUT, THRESHOLD_PROCESSING
     }};
 use crate::msg::ExecuteMsg;
